+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/capability/impl/desktop/ChromeCapabilities.java
@@ -69,7 +69,7 @@ public class ChromeCapabilities extends AbstractCapabilities<ChromeOptions> {
 
         if (Configuration.getBoolean(Configuration.Parameter.AUTO_DOWNLOAD)) {
             chromePrefs.put("download.prompt_for_download", false);
-            if (!"zebrunner".equalsIgnoreCase(R.CONFIG.get("capabilities.provider"))) {
+            if (!"zebrunner".equalsIgnoreCase(R.CONFIG.get(SpecialKeywords.PROVIDER))) {
                 // don't override auto download dir for Zebrunner Selenium Grid (Selenoid)
                 chromePrefs.put("download.default_directory", ReportContext.getArtifactsFolder().getAbsolutePath());
             }
